requirements.txt
# Core LangChain
# langchain==0.1.20
# langchain-community==0.0.38
# langchain-core==0.1.52

# Embeddings & Models
sentence-transformers==2.2.2
transformers==4.40.2

# Core dependencies (pre-built wheels available)
# numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.2

# Web framework
flask==3.0.3
python-dotenv==1.0.1

# PDF & document processing
pypdf==4.2.0

# Vector store & integrations
langchain-pinecone==0.1.1
langchain-openai==0.1.7
langchain-google-genai==1.0.3

# Additional dependencies that might be needed
pinecone-client==3.2.2
openai==1.30.1
google-generativeai==0.5.4

# Text processing
tiktoken==0.7.0